Add EventReview controller, service and dtos

+++ src/main/java/com/example/eventplanner/repositories/event/EventReviewRepository.java
@@ -3,5 +3,5 @@ package com.example.eventplanner.repositories.event;
 import com.example.eventplanner.model.event.EventReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventReviewRepository extends JpaRepository<EventReview, Integer> {
+public interface EventReviewRepository extends JpaRepository<EventReview, Long> {
 }
